new file for debezium related exportData functions
